requirements-dev.txt
-e .
pytest
fakeredis
